Refs #744 moved unit tests under directory hierarchy
