+++ tests/unit/pyrs/core/test_monosetting.py
