+++ tests/unit/pyrs/peaks/test_peak_collection.py
