--- tests/unit/test_polefigurecal.py
@@ -1,27 +1,0 @@
-#!/usr/bin/python
-from pyrs.core import pyrscore
-import pytest
-
-
-def broken_test_pole_figure_calculation():
-    """
-    main testing body to test the workflow to calculate pole figure
-    :return:
-    """
-    # initialize core
-    rs_core = pyrscore.PyRsCore()
-    assert rs_core
-
-    # import data file: detector ID and file name
-    test_data_set = [(1, 'tests/testdata/HB2B_exp129_Long_Al_222[1]_single.hdf5'),
-                     (2, 'tests/testdata/HB2B_exp129_Long_Al_222[2]_single.hdf5'),
-                     (3, 'tests/testdata/HB2B_exp129_Long_Al_222[3]_single.hdf5'),
-                     (4, 'tests/testdata/HB2B_exp129_Long_Al_222[4]_single.hdf5'),
-                     (5, 'tests/testdata/HB2B_exp129_Long_Al_222[5]_single.hdf5'),
-                     (6, 'tests/testdata/HB2B_exp129_Long_Al_222[6]_single.hdf5'),
-                     (7, 'tests/testdata/HB2B_exp129_Long_Al_222[7]_single.hdf5')]
-    assert test_data_set
-
-
-if __name__ == '__main__':
-    pytest.main()
